Split oversized scheduled task and upload modules
