+++ frontend/src/components/panels/ScheduledTaskPanel/SessionScheduledTasksButton.tsx
@@ -1,0 +1,280 @@
+import { useCallback, useEffect, useMemo, useState } from "react";
+import { useTranslation } from "react-i18next";
+import { useNavigate } from "react-router-dom";
+import toast from "react-hot-toast";
+import { Bot, CalendarClock, Clock, History, Loader2, Pause, Play, RotateCcw, Timer } from "lucide-react";
+import { scheduledTaskApi } from "../../../services/api/scheduledTask";
+import { useAuth } from "../../../hooks/useAuth";
+import { Permission } from "../../../types";
+import type { ScheduledTask } from "../../../types/scheduledTask";
+import { formatDateTimeShort } from "../../../utils/datetime";
+import { closePersistentToolPanel, isPersistentToolPanelOpen, openPersistentToolPanel, updatePersistentToolPanel } from "../../chat/ChatMessage/items/persistentToolPanelState";
+import { RunStatusBadge, StatusBadge } from "./Badges";
+import { formatTaskTrigger } from "./utils";
+
+const SESSION_TASK_PANEL_KEY = "session-scheduled-tasks";
+
+function SessionScheduledTaskPanelBody({
+  sessionId,
+  refreshKey,
+  onMutate,
+}: {
+  sessionId: string;
+  refreshKey?: string | number;
+  onMutate?: () => void;
+}) {
+  const { t } = useTranslation();
+  const navigate = useNavigate();
+  const { hasPermission } = useAuth();
+  const canWrite = hasPermission(Permission.SCHEDULED_TASK_WRITE);
+  const [tasks, setTasks] = useState<ScheduledTask[]>([]);
+  const [isLoading, setIsLoading] = useState(true);
+
+  const fetchTasks = useCallback(async () => {
+    setIsLoading(true);
+    try {
+      const response = await scheduledTaskApi.listBySession(sessionId, 0, 50);
+      setTasks(response.items);
+    } catch (error) {
+      const message =
+        error instanceof Error ? error.message : t("common.loadFailed");
+      toast.error(message);
+    } finally {
+      setIsLoading(false);
+    }
+  }, [sessionId, t]);
+
+  useEffect(() => {
+    void fetchTasks();
+  }, [fetchTasks, refreshKey]);
+
+  const runTaskAction = async (
+    task: ScheduledTask,
+    action: "pause" | "resume" | "runNow",
+  ) => {
+    if (!canWrite) {
+      toast.error(t("errors.noPermission"));
+      return;
+    }
+
+    try {
+      if (action === "pause") {
+        await scheduledTaskApi.pause(task.id);
+        toast.success(t("scheduledTask.pausedSuccess"));
+      } else if (action === "resume") {
+        await scheduledTaskApi.resume(task.id);
+        toast.success(t("scheduledTask.resumedSuccess"));
+      } else {
+        await scheduledTaskApi.runNow(task.id);
+        toast.success(t("scheduledTask.triggeredSuccess"));
+      }
+      await fetchTasks();
+      onMutate?.();
+    } catch (error) {
+      const message =
+        error instanceof Error ? error.message : t("common.operationFailed");
+      toast.error(message);
+    }
+  };
+
+  return (
+    <div className="flex h-full min-h-0 flex-col">
+      {isLoading && tasks.length === 0 ? (
+        <div className="flex flex-1 items-center justify-center">
+          <Loader2 className="h-5 w-5 animate-spin text-stone-400" />
+        </div>
+      ) : tasks.length === 0 ? (
+        <div className="flex flex-1 flex-col items-center justify-center px-6 text-center">
+          <CalendarClock className="mb-3 h-9 w-9 text-stone-300 dark:text-stone-600" />
+          <p className="text-sm text-stone-500 dark:text-stone-400">
+            {t(
+              "scheduledTask.noConversationTasks",
+              "当前会话暂无 agent 创建的定时任务",
+            )}
+          </p>
+        </div>
+      ) : (
+        <div className="flex-1 space-y-2 overflow-y-auto p-3">
+          {tasks.map((task) => (
+            <div
+              key={task.id}
+              className="rounded-lg border border-[var(--theme-border)] bg-[var(--theme-bg-card)] p-3"
+            >
+              <div className="flex items-start justify-between gap-2">
+                <div className="min-w-0 flex-1">
+                  <div className="flex items-center gap-2">
+                    <p className="truncate text-sm font-medium text-stone-900 dark:text-stone-100">
+                      {task.name}
+                    </p>
+                    <StatusBadge status={task.status} />
+                  </div>
+                  {task.description && (
+                    <p className="mt-1 line-clamp-2 text-xs text-stone-500 dark:text-stone-400">
+                      {task.description}
+                    </p>
+                  )}
+                </div>
+              </div>
+
+              <div className="mt-3 space-y-1 text-xs text-stone-500 dark:text-stone-400">
+                <div className="flex items-center gap-1.5">
+                  <Timer size={12} />
+                  <span className="truncate">{formatTaskTrigger(task, t)}</span>
+                </div>
+                <div className="flex items-center gap-1.5">
+                  <Bot size={12} />
+                  <span className="truncate">{task.agent_id}</span>
+                </div>
+                {task.last_run_at ? (
+                  <div className="flex items-center gap-1.5">
+                    <Clock size={12} />
+                    <span>{formatDateTimeShort(task.last_run_at)}</span>
+                    {task.last_run_status && (
+                      <RunStatusBadge status={task.last_run_status} />
+                    )}
+                  </div>
+                ) : (
+                  <div>{t("scheduledTask.neverRun")}</div>
+                )}
+              </div>
+
+              <div className="mt-3 flex items-center justify-between gap-2">
+                <button
+                  onClick={() => {
+                    closePersistentToolPanel();
+                    const params = new URLSearchParams({
+                      taskId: task.id,
+                      taskName: task.name,
+                    });
+                    navigate(
+                      `/scheduled-tasks?${params.toString()}`,
+                    );
+                  }}
+                  className="inline-flex h-8 items-center gap-1.5 rounded-lg px-2 text-xs text-stone-500 transition-colors hover:bg-stone-100 hover:text-stone-700 dark:text-stone-400 dark:hover:bg-stone-800 dark:hover:text-stone-200"
+                >
+                  <History size={14} />
+                  {t("scheduledTask.details", "详情")}
+                </button>
+
+                <div className="flex items-center gap-1">
+                  {canWrite && task.status === "active" && (
+                    <button
+                      onClick={() => void runTaskAction(task, "pause")}
+                      className="flex h-8 w-8 items-center justify-center rounded-lg text-stone-400 transition-colors hover:bg-stone-100 hover:text-stone-600 dark:hover:bg-stone-800 dark:hover:text-stone-300"
+                      title={t("scheduledTask.pause")}
+                    >
+                      <Pause size={15} />
+                    </button>
+                  )}
+                  {canWrite && task.status === "paused" && (
+                    <button
+                      onClick={() => void runTaskAction(task, "resume")}
+                      className="flex h-8 w-8 items-center justify-center rounded-lg text-emerald-500 transition-colors hover:bg-emerald-50 dark:hover:bg-emerald-900/30"
+                      title={t("scheduledTask.resume")}
+                    >
+                      <Play size={15} />
+                    </button>
+                  )}
+                  {canWrite && (
+                    <button
+                      onClick={() => void runTaskAction(task, "runNow")}
+                      className="flex h-8 w-8 items-center justify-center rounded-lg text-stone-400 transition-colors hover:bg-blue-50 hover:text-blue-500 dark:hover:bg-blue-900/30 dark:hover:text-blue-400"
+                      title={t("scheduledTask.runNow")}
+                    >
+                      <RotateCcw size={15} />
+                    </button>
+                  )}
+                </div>
+              </div>
+            </div>
+          ))}
+        </div>
+      )}
+    </div>
+  );
+}
+
+export function SessionScheduledTasksButton({
+  sessionId,
+  refreshKey,
+}: {
+  sessionId: string | null;
+  refreshKey?: string | number;
+}) {
+  const { t } = useTranslation();
+  const { hasPermission } = useAuth();
+  const canRead = hasPermission(Permission.SCHEDULED_TASK_READ);
+  const [count, setCount] = useState(0);
+
+  const fetchCount = useCallback(async () => {
+    if (!sessionId || !canRead) {
+      setCount(0);
+      return;
+    }
+    try {
+      const response = await scheduledTaskApi.listBySession(sessionId, 0, 1);
+      setCount(response.total);
+    } catch {
+      setCount(0);
+    }
+  }, [canRead, sessionId]);
+
+  useEffect(() => {
+    void fetchCount();
+  }, [fetchCount, refreshKey]);
+
+  const panelContent = useMemo(() => {
+    if (!sessionId) return null;
+    return (
+      <SessionScheduledTaskPanelBody
+        sessionId={sessionId}
+        refreshKey={refreshKey}
+        onMutate={fetchCount}
+      />
+    );
+  }, [fetchCount, refreshKey, sessionId]);
+
+  useEffect(() => {
+    if (!panelContent || !isPersistentToolPanelOpen(SESSION_TASK_PANEL_KEY)) {
+      return;
+    }
+    updatePersistentToolPanel(
+      (panel) => ({
+        ...panel,
+        children: panelContent,
+        subtitle: count > 0 ? `${count}` : undefined,
+      }),
+      SESSION_TASK_PANEL_KEY,
+    );
+  }, [count, panelContent]);
+
+  if (!sessionId || !canRead || count === 0) return null;
+
+  const togglePanel = () => {
+    if (isPersistentToolPanelOpen(SESSION_TASK_PANEL_KEY)) {
+      closePersistentToolPanel();
+      return;
+    }
+    openPersistentToolPanel({
+      title: t("scheduledTask.conversationTasks", "会话定时任务"),
+      icon: <CalendarClock size={16} />,
+      status: "idle",
+      subtitle: `${count}`,
+      panelKey: SESSION_TASK_PANEL_KEY,
+      children: panelContent,
+    });
+  };
+
+  return (
+    <button
+      onClick={togglePanel}
+      className="absolute right-3 top-3 z-40 flex h-9 w-9 items-center justify-center rounded-full border border-[var(--theme-border)] bg-[var(--theme-bg-card)]/90 text-stone-500 shadow-sm backdrop-blur transition-colors hover:bg-stone-100 hover:text-stone-700 dark:text-stone-300 dark:hover:bg-stone-800"
+      title={t("scheduledTask.conversationTasks", "会话定时任务")}
+    >
+      <CalendarClock size={17} />
+      <span className="absolute -right-1 -top-1 flex h-5 min-w-[20px] items-center justify-center rounded-full bg-orange-500 px-1 text-[10px] font-semibold leading-none text-white">
+        {count > 99 ? "99+" : count}
+      </span>
+    </button>
+  );
+}
