+++ frontend/src/components/panels/ScheduledTaskPanel/types.ts
@@ -1,0 +1,11 @@
+export interface RunConversationMessage {
+  role: "user" | "assistant";
+  content: string;
+  timestamp: string;
+}
+
+export interface ScheduledTaskDefaults {
+  agentId?: string;
+  modelId?: string;
+  modelValue?: string;
+}
